Add MaixCam2 Thermal160 reset and history recorder
