+++ projects/app_thermal160_camera/app.yaml
@@ -4,8 +4,8 @@ name[zh]: 热成像仪160
 version: 1.0.0
 icon: assets/thermal.json
 author: Sipeed Ltd
-desc: Thermal Camera
-desc[zh]: 热成像仪
+desc: Thermal Camera with history recorder
+desc[zh]: 热成像仪和历史记录
 include:
   - assets/thermal.json
   - app.yaml
