--- projects/maix-thermal160_camera_history/app.yaml
@@ -1,18 +1,0 @@
-id: thermal160_camera_history
-name: Thermal160 History
-name[zh]: 热成像160历史记录
-version: 1.0.0
-icon: assets/thermal.json
-author: Sipeed Ltd
-desc: 24-hour Thermal160 temperature history recorder
-desc[zh]: Thermal160 24小时温度历史记录
-include:
-  - assets/thermal.json
-  - app.yaml
-  - main.py
-files:
-  - assets\thermal.json
-  - app.yaml
-  - main.py
-  - README_EN.md
-  - README.md
